requirements.txt
Cython==3.0.2
librosa==0.10.1
matplotlib==3.7.2
# numpy==1.24.4
scipy==1.11.2
tensorboard
onnx==1.14.1
onnxruntime==1.15.1
vector_quantize_pytorch
vinorm
underthesea
eng_to_ipa
g2p_en
